Refactor theme
- Split into separate files
- Update createTheme to allow entire theme to be overridden,
  but still based on our defaults.
- Allow props & overrides to be a function with the compiled theme sans those.
